3.0 switch deafault template for frontend login

+++ src/classes/frontend-login-plus.php
@@ -901,9 +901,9 @@ class FrontendLoginPlus extends Config implements RequiredFunctions {
 				$page_template = apply_filters( 'uo_login_ui_template', $page_template );
 				break;*/
 			default:
-				$page_template = self::get_template( '/login-page-ui-default.php' );
+				$page_template = self::get_template( '/login-page-ui.php' );
 				// TODO var dump this to test
-				$page_template = apply_filters( 'uo_login_ui_template', $page_template );
+				//$page_template = apply_filters( 'uo_login_ui_template', $page_template );
 				break;
 		}
 
@@ -1295,16 +1295,18 @@ class FrontendLoginPlus extends Config implements RequiredFunctions {
 				$login_page_id = $setting['value'];
 				if ( $post->ID == (int) $login_page_id ) {
 					if ( ! has_shortcode( $post->post_content, 'uo_login_ui' ) ) {
-						$blocks           = parse_blocks( $post->post_content );
 						$block_is_on_page = false;
-						foreach ( $blocks as $block ) {
-							if ( 'uncanny-toolkit/login-uncanny' === $block['blockName'] ) {
-								$block_is_on_page = true;
+						if( function_exists('parse_blocks')){
+							$blocks           = parse_blocks( $post->post_content );
+							foreach ( $blocks as $block ) {
+								if ( 'uncanny-toolkit/login-uncanny' === $block['blockName'] ) {
+									$block_is_on_page = true;
+								}
 							}
-						}
-						if ( ! $block_is_on_page ) {
-							echo '<div id="ult-login-no-setup-notice"><strong>Note: This page has been set as the login page for this site.  The form below has been added for your convenience.  To hide this message, add the shortcode [uo_login_ui] to this page.</strong></div>';
-							echo do_shortcode( '[uo_login_ui]' );
+							if ( ! $block_is_on_page ) {
+								echo '<div id="ult-login-no-setup-notice"><strong>Note: This page has been set as the login page for this site.  The form below has been added for your convenience.  To hide this message, add the shortcode [uo_login_ui] to this page.</strong></div>';
+								echo do_shortcode( '[uo_login_ui]' );
+							}
 						}
 					}
 				}
